# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daymark memory app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
        '500':
          description: Failed to generate upload URL.
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve a private object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
...
  /people:
    get:
      operationId: listPeople
      tags: [people]
      summary: List all people
      responses:
        "200":
          description: List of people
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
    post:
      operationId: createPerson
      tags: [people]
      summary: Add a person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonInput"
      responses:
        "201":
          description: Created person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /people/{id}:
    get:
      operationId: getPerson
      tags: [people]
      summary: Get a person and their shared memories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Person detail with memories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Memory:
      type: object
      required: [id, title, date, category, giftColor, ribbon, isKeptClose, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        story:
          type: ["string", "null"]
        date:
          type: string
          format: date
        location:
          type: ["string", "null"]
        category:
          type: string
          enum: [travel, friends, family, achievements, everyday]
        mood:
          type: ["string", "null"]
        giftColor:
          type: string
        ribbon:
          type: string
        sticker:
          type: ["string", "null"]
        photoUrls:
          type: array
          items:
            type: string
        people:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        isKeptClose:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Person:
      type: object
      required: [id, name]
      properties:
        id:
          type: number
        name:
          type: string
        relationship:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
          format: date
        memoriesCount:
          type: number
    PersonDetail:
      type: object
      required: [id, name, memories]
      properties:
        id:
          type: number
        name:
          type: string
        relationship:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
          format: date
        memoriesCount:
          type: number
        nextImportantDate:
          type: ["string", "null"]
        memories:
          type: array
          items:
            $ref: "#/components/schemas/Memory"
    PersonInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        relationship:
          type: string
        avatarUrl:
          type: string
        birthday:
          type: string
          format: date
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: number
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
...